backend/requirements/base.txt
# Django
django==5.2.3
django-environ==0.12.0

# Django REST Framework
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-yasg==1.21.10
